py_redditor/requirements.txt
requests
pandas
praw
python-dotenv